Finished FindAllReposSince method
This method will search through each page of the site until it finds a
repo that is older than the provided date. It will also filter out repos
in the blacklist, and makes sure to wait the delay time between
requests.

Next problem is that it needs to be authenticated or else it can't crawl
the page at any reasonable rate.